pyproject.toml
[tool.poetry]
name = "htmltopdf"
version = "0.1.0"
description = ""
authors = ["Vadim Lakhterman <vadim.l@labos.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic-settings = "^2.4.0"
pdfkit = "^1.0.0"
chardet = "^5.2.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
